Make SongList component
